+++ KSamplerInputs.py
@@ -2,12 +2,6 @@ import os
 from datetime import datetime
 from sys import float_info
 import json
-import piexif
-import piexif.helper
-from PIL import Image
-from PIL.PngImagePlugin import PngInfo
-import numpy as np
-import torch
 import comfy.sd
 
 class AnyType(str):